use forked ParseUint which doesn't require allocating a string
strconv.Parse[U]int requires a string argument.  Because we have a
byte slice, a conversion was required to provide a string (which
allocates a new piece of memory to store an immutable copy of our byte
slice).  This is somewhat silly - even though copying and pasting code
is 'bad', thousands of allocations are worse (6 allocations + memcpy
for each entry in each smaps file).